Fixup some input code. Make the dev console visible when visualing the depth buffer. Add indicators showing where typing occurs in the dev console. Make the AABB visualization mode a console variable that can be toggled.

+++ Win32PlatformIntrinsics.cpp
@@ -386,7 +386,7 @@ void Aligned_Vec4Homogenize(float* data, size_t stride, size_t length) {
   }
 }
 
-void Unaligned_BlendBuffers(float* devBuffer, float* frameBuffer, size_t width, size_t height, float opacity) {
+void Unaligned_BlendBuffers(uint32* devBuffer, uint32* frameBuffer, size_t width, size_t height, float opacity) {
   __m128i devOpacity = _mm_set1_epi16((short)(255.f * opacity));
   __m128i bufferOpacity = _mm_set1_epi16((short)(255.f * (1.f - opacity)));
